+ working on operator evaluation (in progress)...
