+++ src/main/scala/eu/inn/parser/eval/EvalEntityNotFound.scala
@@ -1,3 +1,3 @@
-package eu.inn.parser
+package eu.inn.parser.eval
 
 class EvalEntityNotFound(entityName: String) extends RuntimeException(s"$entityName isn't found")
