+++ src/main/scala/eu/inn/parser/eval/EvalOperationIsNotApplicable.scala
@@ -1,0 +1,3 @@
+package eu.inn.parser.eval
+
+class EvalOperationIsNotApplicable(operation: String) extends RuntimeException(s"$operation is not applicable")
